+++ bin/proving-service/src/commands/proxy.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use clap::Parser;
 use pingora::{
     apps::HttpServerOptions,
@@ -14,7 +16,7 @@ use crate::{
     commands::PROXY_HOST,
     error::ProvingServiceError,
     proxy::{
-        LoadBalancer, LoadBalancerState, status::ProxyStatusService,
+        LoadBalancer, LoadBalancerState, status::ProxyStatusPingoraService,
         update_workers::LoadBalancerUpdateService,
     },
     utils::check_port_availability,
@@ -124,11 +126,13 @@ impl StartProxy {
             info!(target: COMPONENT, "Metrics service disabled");
         }
 
-        // Add status service
-        let status_service = ProxyStatusService::new(worker_lb);
-        let mut status_service = Service::new("status".to_string(), status_service);
-        status_service
-            .add_tcp(format!("{}:{}", PROXY_HOST, self.proxy_config.status_port).as_str());
+        // Add gRPC status service
+        let status_service = ProxyStatusPingoraService::new(
+            worker_lb,
+            self.proxy_config.status_port,
+            Duration::from_secs(self.proxy_config.status_update_interval_secs),
+        )
+        .await;
         info!(target: COMPONENT,
             endpoint = %format!("{}:{}/status", PROXY_HOST, self.proxy_config.status_port),
             "Status service initialized"